requirements.txt
Pillow>=9.0
requests>=2.28
pytz>=2023.0
waveshare-epaper>=1.4.0
# RPi.GPIO and spidev are Pi-only; install.sh handles these
